Rename init_read_declarations to forward_declarations (preprocessor)
