+++ src/stub/preprocessor.rs
@@ -1,4 +1,4 @@
-pub mod init_read_declarations;
+pub mod forward_declarations;
 pub mod lisp_like;
 
 use dyn_clone::DynClone;
