+++ src/stub/preprocessor/forward_declarations.rs
